+++ CHANGELOG.md
@@ -19,7 +19,7 @@
     + NSString methods: `-numberValue`, `-dataValue`
     + `NSURLComponents (ESAdditions)` category
     + `-[NSURL URLByAddingQueryDictionary:]`
-    + NSData methods: `-uppercaseHexString`, `lowercaseHexString`
+    + NSData methods: `+dataWithHexString:`, `-uppercaseHexString`, `-lowercaseHexString`
     + NSDate methods: `+dateFromHTTPDateString:`, `-isToday`, `-isYesterday`, `-isTomorrow`, `-isWeekend`, `-isWorkday`, `-isInThisWeek`
     + NSDateFormatter additions: `+RFC1123DateFormatter`, `+RFC1036DateFormatter`, `+ANSIDateFormatter`
     + NSArray methods: `-objectOrNilAtIndex:`, `-reversedArray`, `-previousObjectToIndex:`, `-previousObjectToObject:`, `-nextObjectToIndex:`, `-nextObjectToObject:`